feat: support to_pretty() and Log.to_pretty()
Log.to_pretty(obj) can be used to get a colorized string version of the
passed object - this is what is eventually printed via print_rich, but
it can be useful for UIs as well (e.g. passed into a RichText Label)

to_pretty() is now a third method that can be implemented on an
object/class/custom_resource to specify the logged data when that object
is logged.

Also refactors and fixes the tests a bit, in preparation for testing
more types.