feat: Log.register_overwrite(type_or_class, func(x, opts))
Adds support for registering a handler for a type or class. The handler
is stored on a Log.log_overwrites static var and used to cut off the
default log implementation for that type/class.

Easier than i thought to implement! Perhaps we could use a similar
structure for log's config data (e.g. max_array_size, color_scheme, etc)