feat: proof of concept for swapping color schemes
Inits a Log.config dictionary that pulls the color scheme from
config.color_scheme. The colors can also be passed to Log.to_pretty in
the opts dict.

Once the schemes are being merged, it'll be possible to overwrite
specific types without clearing the rest of the colors.

Ideally the prettier colors would be opt-ed into by the editor-launched
games, and the term-safe would be used in other circumstances (test
runners, 'built' production games)

A similar feature should make it easy to opt-out of colors - I wonder if
it's easy to read a `--log-no-colors` arg from the command line.