feat: support color overwriting via Log.to_pretty
The color schemes are now built once in Log.to_pretty, and merged into a
single map. This lets a minimal amount of colors be overwritten by a
passed dictionary, and leaves the rest of the colors as-is.

Passing color overwrites at the log callsite is a pain at the moment -
the variadic log functions don't make it clear where an opts dictionary
could be passed, and passing the result of Log.pretty to one of the
public Log funcs will print it as a string, which is not ideal