refactor: move config from editor- to projectSettings
EditorSettings are per-machine and not accessible when starting
scenes (even from the editor). So instead we use the projectSettings,
to at least get these config vals going.

This might make more sense than editor settings, and could be useful to
override with feature tags and what not.

Still a bit of a mess, naming-wise, and seems to throw errors before the
settings have been created in the project :/