fix: resolve type warnings in test file
Also disables a few more warnings - no need to handle return values in
test cases, and exclude_addons for now so we don't see a million gdunit
type warnings.

It should be noted that exclude_addons should be toggled while working
on log.gd to see the errors in that file.