ci: drop godot 4.2 and 4.3
The 4.3 tests are failing b/c vectors print as 1.0 vs 1 between the
versions.

The 4.2 tests are failing b/c my gdunit version itself isn't 4.2
compat (for whatever reason).

Log.gd likely works fine on these versions, but for now I just want the
green checkmark - if there is a need to restore these, we can do so, but
most likely folks are moving to 4.4 and 4.5 these days anyway.