feat: Force Termsafe Colors config option
Adds a boolean config option that will opt-in to termsafe colors in
rebuild_config.

This lets users use term-safe colors without needing a script/autoload
to call `Log.use_termsafe_colors()`.

This used to be possible prior to the color-resource theme feature - in
the future we might consider a set of term-safe LogColorThemes to allow
folks to customize the termsafe colors as well.